Autentification view separated
1. Autentification view separated to a different views(SignInView and SignUpView) and managed in AuthView.
2. Added AuthViewModel for AuthView.
