+++ ReFix/Views/Modifiers/ClearButtonViewModifier.swift
